moved to sub-modules and made CLI work
