+++ .swiftlint.yml
@@ -7,3 +7,5 @@ excluded: # paths to ignore during linting. Takes precedence over `included`.
   - SwiftyJSONAcceleratorTests
   - Pods
   - SwiftyJSONAccelerator/External-Libraries
+  - SwiftyJSON
+  - SwiftCLI
