+++ SwiftyJSONAccelerator/Helpers/JSON+Helpers.swift
@@ -7,7 +7,6 @@
 //
 
 import Foundation
-import SwiftyJSON
 
 // Helper methods for JSON.
 extension JSON {
