+++ SwiftyJSONAccelerator/Library-Extensions/ObjectMapperModelFile.swift
@@ -7,7 +7,6 @@
 //
 
 import Foundation
-import SwiftyJSON
 
 struct ObjectMapperModelFile: ModelFile, DefaultModelFileComponent {
