+++ SwiftyJSONAccelerator/Library-Extensions/SwiftyJSONModelFile.swift
@@ -7,7 +7,6 @@
 //
 
 import Foundation
-import SwiftyJSON
 
 /**
  *  Provides support for SwiftyJSON library.
