--- SwiftyJSONAcceleratorCLI/main.swift
@@ -1,13 +1,0 @@
-//
-//  main.swift
-//  SwiftyJSONAcceleratorCLI
-//
-//  Created by Karthik on 22/12/2016.
-//  Copyright © 2016 Karthikeya Udupa K M. All rights reserved.
-//
-
-import CoreServices
-import Foundation
-
-let instance = SJA()
-instance.staticMode()
